src/main/java/edu.matc/persistence/UserDao.java
package edu.matc.persistence;

import edu.matc.entity.User;
import org.apache.log4j.Logger;
import org.hibernate.*;
import org.hibernate.criterion.Restrictions;

import java.util.List;

public class UserDao {

    private final Logger log = Logger.getLogger(this.getClass());



    /** Return a list of all clients
     *
     * @return All clients
     */
    public List<User> getAllUsers() {
        Session session = SessionFactoryProvider.getSessionFactory().openSession();
        List<User> users = session.createCriteria(User.class).list();
        session.close();
        return users;
    }



    /**
     * retrieve a client given their id
     *
     * @param id the client's id
     * @return client
     */
    public User getUser(int id) {
        Session session = SessionFactoryProvider.getSessionFactory().openSession();
        Transaction tx = null;
        User user = null;
        try {
            tx = session.beginTransaction();
            user = (User) session.get(User.class, id);
            tx.commit();
        } catch (HibernateException he) {
            if (tx!=null) {
                tx.rollback();
                log.error("Error retrieving user id: " + id, he);
            }
        } finally {
            session.close();
        }

        return user;
    }

}
